added new rewinding reflection bindings. still need to implement rewinding callback and do stuff with EventRewind. did some bugfixing related to codegen aswell as a feature update (to support new binding creation).

+++ codegen/reflection/objects/ReflectableObjects.py
@@ -61,7 +61,7 @@ class TeamData(ReflectableObject):
     orderCooldownTotal = Var("uint32_t", 4)
     orderCooldownLeft = Var("uint32_t", 5)
     spawnScore = Var("uint32_t", 6)
-    roundScore = Var("uint32_t", 7)
+    roundScore = Var("float", 7)
 
 class GlobalElo(ReflectableObject):
     teamAvgEloRatings = Var("danet::teamAvgEloRatings", 2)
