+++ replay/include/mpi/codegen/ReflectableObjects/TeamData.h
@@ -9,7 +9,7 @@ public:
   danet::ReflectionVar<uint32_t> orderCooldownTotal{"orderCooldownTotal", &orderCooldownLeft, 4};
   danet::ReflectionVar<uint32_t> orderCooldownLeft{"orderCooldownLeft", &spawnScore, 5};
   danet::ReflectionVar<uint32_t> spawnScore{"spawnScore", &roundScore, 6};
-  danet::ReflectionVar<uint32_t> roundScore{"roundScore", nullptr, 7};
+  danet::ReflectionVar<float> roundScore{"roundScore", nullptr, 7};
   TeamData() : ReflectableObject()  {
     varList.head = &score;
     varList.tail = &roundScore;
